Document how to run on a cluster with spark-submit.
This change creates a Spark application JAR that can be used to
run the word count example on the cluster.  The Spark depdendency
is now marked 'provided' to make it possible to run on different
(minor) versions of Spark.

Note also that Guava is shaded in the application JAR to avoid
conflicts with the version used by Hadoop.