Allow tests to share a single Spark context.
It's not possible to have multiple Spark contexts per JVM (even if not running concurrently),
and while one solution is to fork a new JVM per test class, it's not always possible or desirable
to do this - if for example the test suite has a dataflow test per test method, or the test
suite has many very small tests.

This JIRA has more background and discussion about the JVM limitation:

https://issues.apache.org/jira/browse/SPARK-2243

This change makes it possible to share a SparkContext when running tests that use the
SparkPipelineRunner by setting the system property 'dataflow.spark.test.reuseSparkContext'
to 'true'.